+++ CHANGELOG.md
@@ -17,12 +17,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - rename columns
 - Converter for Lithuania (EuroCrops)
 - Converter for Switzerland
+- `fiboa convert`: New parameter `--original-geometries` / `-og` to keep the original geometries
 
 ### Changed
 
-- `fiboa convert` writes custom schemas to collection metadata
+- `fiboa convert`:
+  - Writes custom schemas to collection metadata
+  - Geometries are made valid using GeoPanda's `make_valid` method by default
+  - MultiPolygons are converted to Polygons by default
 - `fiboa validate` uses custom schemas for validation
 - `fiboa merge` keeps custom schemas when needed
+
+### Removed
+- `fiboa convert`: Removed the explicit parameter `explode_multipolygon` from the converters
 
 ### Fixed